Renamed `jt-component` to `component`
This is because there is currently no reason to give the component library such a custom name.
