+++ lib/component/element-map.js
