+++ lib/component/main.js
