+++ lib/component/template.js
