+++ lib/countdown-timer/countdown-timer.js
@@ -1,6 +1,6 @@
 /* globals HTMLElement, CustomEvent */
 
-import { Template } from '../jt-component/main.js'
+import { Template } from '../component/main.js'
 
 /**
  * Custom component which creates a countdown timer. The following attributes
