+++ lib/editable-list/editable-list.js
@@ -1,6 +1,6 @@
 /* globals HTMLElement */
 
-import { Template } from '../jt-component/main.js'
+import { Template } from '../component/main.js'
 
 /**
  * Custom component which creates a list to which the user can add or remove
